API

FDM.FDMReport — Type.

FDMReport

Details of a finite difference method to estimate a derivative. Instances of FDMReport Base.show nicely.

Fields

p::Int: Order of the method.

q::Int: Order of the derivative that is estimated.

grid::Vector{<:Real}: Relative spacing of samples of f that are used by the method.

coefs::Vector{<:Real}: Weights of the samples of f.

ε::Real: Absolute roundoff error of the function evaluations.

M::Real: Assumed upper bound of f and all its derivatives at x.

ĥ::Real: Step size.

err::Real: Estimated absolute accuracy.

source

FDM.assert_approx_equal — Method.

assert_approx_equal(x, y, ε_abs, ε_rel[, desc])

Assert that x is approximately equal to y.

Let eps_z = eps_abs / eps_rel. Call x and y small if abs(x) < eps_z and abs(y) < eps_z, and call x and y large otherwise. If this function returns True, then it is guaranteed that abs(x - y) < 2 eps_rel max(abs(x), abs(y)) if x and y are large, and abs(x - y) < 2 eps_abs if x and y are small.

Arguments

x: First object to compare.

y: Second object to compare.

ε_abs: Absolute tolerance.

ε_rel: Relative tolerance.

desc: Description of the comparison. Omit or set to false to have no description.

source

FDM.backward_fdm — Method.

backward_fdm(p::Int, ...)

Construct a backward finite difference method of order p. See fdm for further details.

Arguments

p::Int: Order of the method.

Further takes, in the following order, the arguments q, ε, M, and report from fdm.

source

FDM.central_fdm — Method.

central_fdm(p::Int, ...)

Construct a central finite difference method of order p. See fdm for further details.

Arguments

p::Int: Order of the method.

Further takes, in the following order, the arguments q, ε, M, and report from fdm.

source

FDM.fdm — Method.

function fdm(
    grid::Vector{<:Real},
    q::Int;
    ε::Real=eps(),
    M::Real=1,
    report::Bool=false
)

Construct a function method(f, x::Real, h::Real=ĥ) that takes in a function f, a point x in the domain of f, and optionally a step size h, and estimates the q'th order derivative of f at x with a length(grid)'th order finite difference method.

Arguments

grid::Vector{<:Real}: Relative spacing of samples of f that are used by the method. The length of grid determines the order of the method.

q::Int: Order of the derivative to estimate. q must be strictly less than the order of the method.

Keywords

ε::Real=eps(): Absolute roundoff error on the function evaluations.

M::Real=1: Upper bound on f and all its derivatives.

report::Bool=false: Also return an instance of FDMReport containing information about the method constructed.

source

FDM.forward_fdm — Method.

forward_fdm(p::Int, ...)

Construct a forward finite difference method of order p. See fdm for further details.

Arguments

p::Int: Order of the method.

Further takes, in the following order, the arguments q, ε, M, and report from fdm.

source